+++ src/components/DaoCreationWizard/FoundersView/founders.container.ts
@@ -1,0 +1,15 @@
+import Founders from "./founders"
+import { connect } from "react-redux"
+
+const mapStateToProps = (state: any) => {
+  return {}
+}
+
+const mapDispatchToProps = (dispatch: any) => {
+  return {}
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Founders)
