+++ src/components/DaoCreationWizard/FoundersView/index.ts
@@ -1,0 +1,3 @@
+import Founders from "./founders.container"
+
+export default Founders
