+++ src/components/DaoCreationWizard/HorizontalStepper/index.ts
@@ -1,0 +1,1 @@
+// TODO
