+++ src/components/DaoCreationWizard/index.tsx
@@ -1,0 +1,1 @@
+// TODO: create horizontal stepper with the different views
