+++ src/components/Home/index.ts
@@ -1,0 +1,3 @@
+import Home from "./home.container"
+
+export default Home
